Seperated component.js into several files for better readability
