+++ index.html
@@ -1,4 +1,4 @@
 <body>
   <poll-component></poll-component>
-  <script src="./component.js"></script>
+  <script src="./frontend/poll-component.js" type="module"></script>
 </body>
